+++ vscode/src/completions/get-current-doc-context.ts
@@ -29,6 +29,7 @@ export interface DocumentContext extends DocumentDependentContext, LinesContext 
      * on every completion update or new virtual completion creation.
      */
     injectedCompletionText?: string
+    positionWithoutInjectedCompletionText?: vscode.Position
 }
 
 export interface DocumentDependentContext {
@@ -210,7 +211,7 @@ export function insertIntoDocContext(params: InsertIntoDocContextParams): Docume
         text: insertText,
     })
 
-    return getDerivedDocContext({
+    const updatedDocContext = getDerivedDocContext({
         languageId,
         position: updatedPosition,
         dynamicMultilineCompletions,
@@ -222,6 +223,12 @@ export function insertIntoDocContext(params: InsertIntoDocContextParams): Docume
             injectedPrefix: null,
         },
     })
+
+    updatedDocContext.positionWithoutInjectedCompletionText =
+        updatedDocContext.positionWithoutInjectedCompletionText || docContext.position
+    updatedDocContext.injectedCompletionText = (docContext.injectedCompletionText || '') + insertText
+
+    return updatedDocContext
 }
 
 export interface LinesContext {