+++ vscode/src/main.ts
@@ -1,7 +1,6 @@
 import * as vscode from 'vscode'
 
 import {
-    ConfigFeaturesSingleton,
     FeatureFlag,
     featureFlagProvider,
     graphqlClient,
@@ -10,7 +9,7 @@ import {
     PromptMixin,
     setLogger,
     type ConfigurationWithAccessToken,
-    isFixupCommand,
+    ConfigFeaturesSingleton,
 } from '@sourcegraph/cody-shared'
 
 import { ChatManager, CodyChatPanelViewType } from './chat/chat-view/ChatManager'
@@ -25,7 +24,7 @@ import {
     type AuthStatus,
 } from './chat/protocol'
 import { CodeActionProvider } from './code-actions/CodeActionProvider'
-import { newCodyCommandArgs, type CodyCommandArgs } from './commands'
+import type { CodyCommandArgs } from './commands/types'
 import { GhostHintDecorator } from './commands/GhostHintDecorator'
 import { createInlineCompletionItemProvider } from './completions/create-inline-completion-item-provider'
 import { getConfiguration, getFullConfig } from './configuration'
@@ -49,9 +48,19 @@ import { createOrUpdateEventLogger, telemetryService } from './services/telemetr
 import { createOrUpdateTelemetryRecorderProvider, telemetryRecorder } from './services/telemetry-v2'
 import { onTextDocumentChange } from './services/utils/codeblock-action-tracker'
 import { parseAllVisibleDocuments, updateParseTreeOnEdit } from './tree-sitter/parse-tree-cache'
+import { executeCodyCommand, setCommandController } from './commands/CommandsController'
+import { newCodyCommandArgs } from './commands/utils/get-commands'
+import type { DefaultCodyCommands } from '@sourcegraph/cody-shared/src/commands/types'
 import type { FixupTask } from './non-stop/FixupTask'
 import { EnterpriseContextFactory } from './context/enterprise-context-factory'
 import { CodyProExpirationNotifications } from './notifications/cody-pro-expiration'
+import {
+    executeExplainCommand,
+    executeTestCommand,
+    executeSmellCommand,
+    executeDocCommand,
+    executeUnitTestCommand,
+} from './commands/execute'
 
 /**
  * Start the extension, watching all relevant configuration and secrets for changes.
@@ -121,7 +130,6 @@ const register = async (
     await configureEventsInfra(initialConfig, isExtensionModeDevOrTest)
 
     const editor = new VSCodeEditor()
-    const commandsController = platform.createCommandsController?.(editor)
 
     // Could we use the `initialConfig` instead?
     const workspaceConfig = vscode.workspace.getConfiguration()
@@ -204,8 +212,7 @@ const register = async (
         enterpriseContextFactory,
         localEmbeddings || null,
         symfRunner || null,
-        guardrails,
-        commandsController
+        guardrails
     )
 
     const ghostHintDecorator = new GhostHintDecorator()
@@ -297,9 +304,12 @@ const register = async (
     // Sync initial auth status
     await chatManager.syncAuthStatus(authProvider.getAuthStatus())
 
+    const commandsManager = platform.createCommandsProvider?.()
+    setCommandController(commandsManager)
+
     // Execute Cody Commands and Cody Custom Commands
     const executeCommand = (
-        commandKey: string,
+        commandKey: DefaultCodyCommands | string,
         args?: Partial<CodyCommandArgs>
     ): Promise<CommandResult | undefined> => {
         return executeCommandUnsafe(commandKey, args).catch(error => {
@@ -310,36 +320,32 @@ const register = async (
             return undefined
         })
     }
+
     const executeCommandUnsafe = async (
-        commandKey: string,
+        id: DefaultCodyCommands | string,
         args?: Partial<CodyCommandArgs>
     ): Promise<CommandResult | undefined> => {
-        logDebug('executeCommand', commandKey, args)
-        const commandArgs = newCodyCommandArgs(args)
-        const runner = await commandsController?.startCommand(commandKey, commandArgs)
-        const command = runner?.command
-        // Stop processing the command if it is not an ask(chat) request
-        if (!command) {
+        const { commands } = await ConfigFeaturesSingleton.getInstance().getConfigFeatures()
+        if (!commands) {
+            void vscode.window.showErrorMessage(
+                'This feature has been disabled by your Sourcegraph site admin.'
+            )
             return undefined
         }
-        if (isFixupCommand(command)) {
-            return {
-                type: 'edit',
-                task: await runner.fixupTask,
-            }
-        }
 
-        const configFeatures = await ConfigFeaturesSingleton.getInstance().getConfigFeatures()
-        const session = await chatManager.executeCommand(
-            runner.command,
-            commandArgs,
-            configFeatures.commands
-        )
-        return {
-            type: 'chat',
-            session,
-        }
+        // Process command with the commands controller
+        return await executeCodyCommand(id, newCodyCommandArgs(args))
     }
+
+    // Register Cody Commands
+    disposables.push(
+        vscode.commands.registerCommand('cody.action.command', (id, a) => executeCommand(id, a)),
+        vscode.commands.registerCommand('cody.command.explain-code', a => executeExplainCommand(a)),
+        vscode.commands.registerCommand('cody.command.generate-tests', a => executeTestCommand(a)),
+        vscode.commands.registerCommand('cody.command.smell-code', a => executeSmellCommand(a)),
+        vscode.commands.registerCommand('cody.command.document-code', a => executeDocCommand(a)),
+        vscode.commands.registerCommand('cody.command.unit-tests', a => executeUnitTestCommand(a)) // behind unstable flag
+    )
 
     const statusBar = createStatusBar()
 
@@ -383,34 +389,6 @@ const register = async (
             vscode.commands.executeCommand('workbench.action.openSettings', {
                 query: '@ext:sourcegraph.cody-ai chat',
             })
-        ),
-        // Cody Commands
-        vscode.commands.registerCommand('cody.action.commands.menu', async () => {
-            await commandsController?.menu('default')
-        }),
-        vscode.commands.registerCommand('cody.action.commands.custom.menu', () =>
-            commandsController?.menu('custom')
-        ),
-        vscode.commands.registerCommand('cody.settings.commands', () =>
-            commandsController?.menu('config')
-        ),
-        vscode.commands.registerCommand('cody.action.commands.exec', async (title, args) =>
-            executeCommand(title, args)
-        ),
-        vscode.commands.registerCommand('cody.command.explain-code', async args =>
-            executeCommand('/explain', args)
-        ),
-        vscode.commands.registerCommand('cody.command.generate-tests', async args =>
-            executeCommand('/test', args)
-        ),
-        vscode.commands.registerCommand('cody.command.unit-tests', async args =>
-            executeCommand('/unit', args)
-        ),
-        vscode.commands.registerCommand('cody.command.document-code', async args =>
-            executeCommand('/doc', args)
-        ),
-        vscode.commands.registerCommand('cody.command.smell-code', async args =>
-            executeCommand('/smell', args)
         ),
 
         // Account links
@@ -614,9 +592,7 @@ const register = async (
         return setupAutocompleteQueue
     }
 
-    const promises: Promise<void>[] = []
-
-    promises.push(setupAutocomplete().catch(() => {}))
+    const autocompleteSetup = setupAutocomplete().catch(() => {})
 
     if (initialConfig.experimentalGuardrails) {
         const guardrailsProvider = new GuardrailsProvider(guardrails, editor)
@@ -627,9 +603,11 @@ const register = async (
         )
     }
 
-    promises.push(showSetupNotification(initialConfig))
-
-    await Promise.all(promises)
+    // INC-267 do NOT await on this promise. This promise triggers
+    // `vscode.window.showInformationMessage()`, which only resolves after the
+    // user has clicked on "Setup". Awaiting on this promise will make the Cody
+    // extension timeout during activation.
+    void showSetupNotification(initialConfig)
 
     // Register a serializer for reviving the chat panel on reload
     if (vscode.window.registerWebviewPanelSerializer) {
@@ -642,6 +620,8 @@ const register = async (
             },
         })
     }
+
+    await autocompleteSetup
 
     return {
         disposable: vscode.Disposable.from(...disposables),